requirements.txt
# Core dependencies
pandas>=2.0.0
pyyaml>=6.0
pytz>=2023.3
openpyxl>=3.1.0
python-dateutil>=2.8.2

# Web framework
flask>=3.0.0
plotly>=5.18.0

# Optional: for better performance
fastparquet>=2023.10.0

# Development
pytest>=7.4.0
